added some stats for ccache

+++ scripts/buildMacOs.fish
@@ -1,4 +1,9 @@
 #!/usr/bin/env fish
+if test "$PARALLELISM" = ""
+    set -xg PARALLELISM 64
+end
+echo "Using parallelism $PARALLELISM"
+
 cd $INNERWORKDIR
 mkdir -p .ccache.mac
 set -x CCACHE_DIR $INNERWORKDIR/.ccache.mac
@@ -8,18 +13,18 @@ end
 ccache -M 100G
 cd $INNERWORKDIR/ArangoDB
 
-echo "Starting build at "(date)" on "(hostname)
-ccache --zero-stats
-
-rm -rf build
+cd $INNERWORKDIR/ArangoDB
+if test -z "$NO_RM_BUILD"
+  echo "Cleaning build directory"
+  rm -rf build
+end
 mkdir -p build
 cd build
 
-echo cmake $argv -DCMAKE_BUILD_TYPE=$BUILDMODE -DCMAKE_CXX_COMPILER=$CCACHEBINPATH/g++ -DCMAKE_C_COMPILER=$CCACHEBINPATH/gcc -DUSE_MAINTAINER_MODE=$MAINTAINER -DUSE_ENTERPRISE=$ENTERPRISEEDITION -DUSE_JEMALLOC=Off -DCMAKE_SKIP_RPATH=On -DPACKAGING=Bundle -DPACKAGE_TARGET_DIR=$INNERWORKDIR -DOPENSSL_USE_STATIC_LIBS=On ..
+echo "Starting build at "(date)" on "(hostname)
+ccache --zero-stats
 
-echo cmake output in $INNERWORKDIR/cmakeArangoDB.log
-
-cmake $argv \
+set -g FULLARGS $argv \
       -DCMAKE_BUILD_TYPE=$BUILDMODE \
       -DCMAKE_CXX_COMPILER=$CCACHEBINPATH/g++ \
       -DCMAKE_C_COMPILER=$CCACHEBINPATH/gcc \
@@ -29,9 +34,26 @@ cmake $argv \
       -DCMAKE_SKIP_RPATH=On \
       -DPACKAGING=Bundle \
       -DPACKAGE_TARGET_DIR=$INNERWORKDIR \
-      -DOPENSSL_USE_STATIC_LIBS=On \
-      .. > $INNERWORKDIR/cmakeArangoDB.log ^&1
-and echo "Finished cmake at "(date)", now starting build"
+      -DOPENSSL_USE_STATIC_LIBS=On
+
+if test "$ASAN" = "On"
+  echo "ASAN is not support in this environment"
+end
+
+echo cmake $FULLARGS ..
+echo cmake output in $INNERWORKDIR/cmakeArangoDB.log
+
+cmake $FULLARGS .. > $INNERWORKDIR/cmakeArangoDB.log ^&1
+or exit $status
+
+echo "Finished cmake at "(date)", now starting build"
+
+set -g MAKEFLAGS -j$PARALLELISM 
+if test "$VERBOSEBUILD" = "On"
+  echo "Building verbosely"
+  set -g MAKEFLAGS $MAKEFLAGS V=1 VERBOSE=1 Verbose=1
+end
+
 and echo Running make, output in $INNERWORKDIR/buildArangoDB.log
 and nice make -j$PARALLELISM > $INNERWORKDIR/buildArangoDB.log ^&1
